Users count on badges computed no the fly

+++ KikoleSite/Api/Badge.cs
@@ -8,7 +8,7 @@
 
         public string Description { get; set; }
 
-        public ulong Users { get; set; }
+        public int Users { get; set; }
 
         public bool Hidden { get; set; }
     }
